Docstrings

This page lists all docstrings in EquivariantTensors.jl including for functions that are not part of the public API. Please check with the Public API which functionality for which we aim to guarantee semver-stability.

EquivariantTensors.O3.coupling_coeffs — Function

O3.coupling_coeffs(L, ll, nn; PI, basis)
O3.coupling_coeffs(L, ll; PI, basis)

Compute coupling coefficients for the spherical harmonics basis, where

L must be an Integer;

ll, nn must be vectors or tuples of Integer of the same length.

PI: whether or not the coupled basis is permutation-invariant (or the

corresponding tensor symmetric); default is true when nn is provided and false when nn is not provided.

basis: which basis is being coupled, default is complex, alternative

choice is real, which is compatible with the SpheriCart.jl convention.