Docstrings

This page lists all docstrings in EquivariantTensors.jl including for functions that are not part of the public API. Please check with the Public API for which functionality we aim to guarantee semver-stability.

EquivariantTensors.PooledSparseProduct

EquivariantTensors.SparseSymmProd

EquivariantTensors.O3.coupling_coeffs

EquivariantTensors._auto_Rnl_spec

EquivariantTensors._auto_Ylm_spec

EquivariantTensors._auto_nnllmm_spec

EquivariantTensors._gensparse

EquivariantTensors._make_idx_AA_spec

EquivariantTensors._make_idx_A_spec

EquivariantTensors.gensparse

EquivariantTensors.invmap

EquivariantTensors.setproduct

EquivariantTensors.sparse_equivariant_tensor

EquivariantTensors.PooledSparseProduct — Type

struct PooledSparseProduct : This implements a fused (tensor) product and pooling operation. Suppose we are given $N$ embeddings $\phi^{(i)}_{k_i}$ then the pooled sparse product generates feature vectors of the form

\[A_{k_1, \dots, k_N} = \sum_{j} \prod_{t = 1}^N \phi^{(t)}_{k_t}(x_j)\]

where $x_j$ are an list of inputs (multi-set).

The canonical example is

\[A_{nlm} = \sum_{j} R_{nl}(r_j, \mu_j) Y_l^m( \hat{\bm r}_j ),\]

where $R_{nl}$ is a radial embedding, possibly depending on a categorical variable $\mu_j$ and $Y_l^m$ are spherical harmonics.

Constructor

PooledSparseProduct(spec)

where spec is a list of $(k_1, \dots, k_N)$ tuples or vectors, or AbstractMatrix where each column specifies such a tuple.

EquivariantTensors.SparseSymmProd — Type

SparseSymmProd : sparse symmetric product with entries stored as tuples. Input is a vector A; each entry of the output vector AA is of the form

\[ {\bm A}_{i_1, \dots, i_N} = \prod_{t = 1}^N A_{i_t}.\]

Constructor

SparseSymmProd(spec)

where spec is a list of tuples or vectors, each of which specifies an AA basis function as described above. For example,

spec = [ (1,), (2,), (1,1), (1,2), (2,2), 
         (1,1,1), (1,1,2), (1,2,2), (2,2,2) ]
basis = SparseSymmProd(spec)         

defines a basis of 9 functions,

\[[ A_1, A_2, A_1^2, A_1 A_2, A_2^2, A_1^3, A_1^2 A_2, A_1 A_2^2, A_2^3 ]\]

EquivariantTensors._auto_Rnl_spec — Method

Takes a list of 𝔸 or 𝔹 specifications (many-body) in the form of

   [  [(n=., l=., m=.), (n=., l=., m=.)], ... ]

and converts it into a list of sorted unique (n=., l=.) named pairs, i.e the specification of the one-body basis.

EquivariantTensors._auto_Ylm_spec — Method

autoYlmspec(mbspec)

takes a list of 𝔸 or 𝔹 specifications (many-body) and return the specification of the Ylm basis functions as a [ (l = ., m = .), ... ] list.

EquivariantTensors._auto_nnllmm_spec — Method

takes an nnll spec and generates a complete list of all possible nnllmm

EquivariantTensors._gensparse — Method

_gensparse : function barrier for gensparse

EquivariantTensors._make_idx_AA_spec — Method

convert readable AA_spec into the internal representation of the AA basis

EquivariantTensors._make_idx_A_spec — Method

convert readable A_spec into the internal representation of the A basis

EquivariantTensors.gensparse — Method

gensparse(...) : utility function to generate high-dimensional sparse grids which are downsets. All arguments are keyword arguments (with defaults):

NU : maximum correlation order

minvv = 0 : minvv[i] gives the minimum value forvv[i]`

maxvv = Inf : maxvv[i] gives the minimum value forvv[i]`

tup2b = vv -> vv :

admissible = _ -> false : determines whether a tuple belongs to the downset

filter = _ -> true : a callable object that returns true of tuple is to be kept and

false otherwise (whether or not it is part of the downset!) This is used, e.g. to enfore conditions such as ∑ lₐ = even or |∑ mₐ| ≦ M

INT = Int : integer type to be used

ordered = false : whether only ordered tuples are produced; ordered tuples

correspond to permutation-invariant basis functions

EquivariantTensors.invmap — Method

  invmap(a::AbstractVector)

Return a dictionary that maps the elements of a to their indices, to accelerate if there are many repeated searches into a.

EquivariantTensors.setproduct — Method

setproduct(A)

Assumes the A is a length-N collection of collections. It returns a P = Matrix{T} where each P[i, :] is a vector of length N with P[i, j] ∈ A[j]. The number of columns of P is the number of such products, i.e. prod(length.(A)).

In constrast with Iterators.product this implementation is type-stable for a priori unknown N.

EquivariantTensors.sparse_equivariant_tensor — Method

function buildsparseace()

EquivariantTensors.O3.coupling_coeffs — Function

O3.coupling_coeffs(L, ll, nn; PI, basis)
O3.coupling_coeffs(L, ll; PI, basis)

Compute coupling coefficients for the spherical harmonics basis, where

L must be an Integer;

ll, nn must be vectors or tuples of Integer of the same length.

PI: whether or not the coupled basis is permutation-invariant (or the

corresponding tensor symmetric); default is true when nn is provided and false when nn is not provided.

basis: which basis is being coupled, default is complex, alternative

choice is real, which is compatible with the SpheriCart.jl convention.